requirements.txt
plaid-python
python-dotenv
sqlite3